add Impact category
